Move proxy/ to new tools/ folder
